+++ xarray/plot/utils.py
@@ -1,9 +1,11 @@
 from __future__ import absolute_import, division, print_function
 
+import itertools
 import textwrap
 import warnings
 
 import numpy as np
+import pandas as pd
 
 from ..core.options import OPTIONS
 from ..core.pycompat import basestring
@@ -386,3 +388,65 @@ def label_from_attrs(da, extra=''):
         units = ''
 
     return '\n'.join(textwrap.wrap(name + extra + units, 30))
+
+
+def _interval_to_mid_points(array):
+    """
+    Helper function which returns an array
+    with the Intervals' mid points.
+    """
+
+    return np.array([x.mid for x in array])
+
+
+def _interval_to_bound_points(array):
+    """
+    Helper function which returns an array
+    with the Intervals' boundaries.
+    """
+
+    array_boundaries = np.array([x.left for x in array])
+    array_boundaries = np.concatenate(
+        (array_boundaries, np.array([array[-1].right])))
+
+    return array_boundaries
+
+
+def _interval_to_double_bound_points(xarray, yarray):
+    """
+    Helper function to deal with a xarray consisting of pd.Intervals. Each
+    interval is replaced with both boundaries. I.e. the length of xarray
+    doubles. yarray is modified so it matches the new shape of xarray.
+    """
+
+    xarray1 = np.array([x.left for x in xarray])
+    xarray2 = np.array([x.right for x in xarray])
+
+    xarray = list(itertools.chain.from_iterable(zip(xarray1, xarray2)))
+    yarray = list(itertools.chain.from_iterable(zip(yarray, yarray)))
+
+    return xarray, yarray
+
+
+def _resolve_intervals_2dplot(val, func_name):
+    """
+    Helper function to replace the values of a coordinate array containing
+    pd.Interval with their mid-points or - for pcolormesh - boundaries which
+    increases length by 1.
+    """
+    label_extra = ''
+    if _valid_other_type(val, [pd.Interval]):
+        if func_name == 'pcolormesh':
+            val = _interval_to_bound_points(val)
+        else:
+            val = _interval_to_mid_points(val)
+            label_extra = '_center'
+
+    return val, label_extra
+
+
+def _valid_other_type(x, types):
+    """
+    Do all elements of x have a type from types?
+    """
+    return all(any(isinstance(el, t) for t in types) for el in np.ravel(x))
